feat(test): refactor + add granular tests
