+++ sdk/multisig/src/actions/createMultisig.ts
@@ -5,30 +5,17 @@ import {
   SendOptions,
   Signer,
   TransactionInstruction,
+  VersionedTransaction,
 } from "@solana/web3.js";
-import { Member, Multisig, PROGRAM_ID, ProgramConfig } from "../generated";
-import { getMultisigPda, getProgramConfigPda, instructions } from "..";
-import { BaseBuilder, BaseBuilderArgs, BuildResult } from "./common";
-import { SquadPermissions, createMembers } from "./members";
-
-interface CreateMultisigActionArgs extends BaseBuilderArgs {
-  /** The number of approvals required to approve transactions */
-  threshold: number;
-  /** The list of members in the multisig, with their associated permissions */
-  members: Member[];
-  /** Optional time lock in seconds */
-  timeLock?: number;
-  /** Optional config authority key that can override consensus for ConfigTransactions */
-  configAuthority?: PublicKey;
-  /** Optional rent collector where completed transaction rent will go after reclaim */
-  rentCollector?: PublicKey;
-  /** Optional program ID (defaults to Solana mainnet-beta/devnet Program ID) */
-  programId?: PublicKey;
-}
-
-interface CreateMultisigResult extends BuildResult {
-  multisigKey: PublicKey;
-}
+import { Multisig, PROGRAM_ID } from "../generated";
+import { createMultisigCore } from "./common/multisig";
+import {
+  BuildTransactionSettings,
+  CreateMultisigActionArgs,
+  CreateMultisigResult,
+  SendSettings,
+} from "./common/types";
+import { BaseBuilder } from "./common/base";
 
 /**
  * Builds an instruction to create a new Multisig, and returns the instruction and `createKey`
@@ -81,7 +68,7 @@ class MultisigBuilder extends BaseBuilder<
   public multisigKey: PublicKey = PublicKey.default;
 
   constructor(args: CreateMultisigActionArgs) {
-    super(args);
+    super(args, { generateCreateKey: true });
   }
 
   protected async build(): Promise<void> {
@@ -104,7 +91,7 @@ class MultisigBuilder extends BaseBuilder<
         rentCollector,
         programId,
       },
-      this.createKey
+      this.createKey!
     );
 
     this.instructions = [...result.instructions];
@@ -113,7 +100,7 @@ class MultisigBuilder extends BaseBuilder<
 
   async getCreateKey(): Promise<Keypair> {
     await this.ensureBuilt();
-    return this.createKey;
+    return this.createKey!;
   }
 
   async getMultisigKey(): Promise<PublicKey> {
@@ -131,69 +118,46 @@ class MultisigBuilder extends BaseBuilder<
     return multisigAccount;
   }
 
-  async sendAndConfirm(settings?: {
-    preInstructions?: TransactionInstruction[] | undefined;
-    postInstructions?: TransactionInstruction[] | undefined;
-    feePayer?: Signer | undefined;
-    signers?: Signer[] | undefined;
-    options?: SendOptions | undefined;
-  }): Promise<string> {
+  async transaction(
+    settings?: BuildTransactionSettings
+  ): Promise<VersionedTransaction> {
     await this.ensureBuilt();
     if (settings?.signers) {
-      settings.signers.push(this.createKey);
+      settings.signers.push(this.createKey!);
     } else {
       settings = {
-        signers: [this.createKey],
+        signers: [this.createKey!],
+        ...settings,
+      };
+    }
+    return await super.transaction(settings);
+  }
+
+  async send(settings?: SendSettings): Promise<string> {
+    await this.ensureBuilt();
+    if (settings?.signers) {
+      settings.signers.push(this.createKey!);
+    } else {
+      settings = {
+        signers: [this.createKey!],
+        ...settings,
+      };
+    }
+    return await super.send(settings);
+  }
+
+  async sendAndConfirm(settings?: SendSettings): Promise<string> {
+    await this.ensureBuilt();
+    if (settings?.signers) {
+      settings.signers.push(this.createKey!);
+    } else {
+      settings = {
+        signers: [this.createKey!],
         ...settings,
       };
     }
     return await super.sendAndConfirm(settings);
   }
-}
-
-export async function createMultisigCore(
-  args: CreateMultisigActionArgs,
-  createKey: Keypair
-): Promise<CreateMultisigResult> {
-  const {
-    connection,
-    creator,
-    threshold,
-    members,
-    timeLock = 0,
-    configAuthority,
-    rentCollector,
-    programId = PROGRAM_ID,
-  } = args;
-
-  const [multisigPda] = getMultisigPda({
-    createKey: createKey.publicKey,
-    programId,
-  });
-  const programConfigPda = getProgramConfigPda({ programId })[0];
-
-  const programConfig = await ProgramConfig.fromAccountAddress(
-    connection,
-    programConfigPda
-  );
-
-  const ix = instructions.multisigCreateV2({
-    creator,
-    threshold,
-    members,
-    multisigPda: multisigPda,
-    treasury: programConfig.treasury,
-    createKey: createKey.publicKey,
-    timeLock: timeLock ?? 0,
-    rentCollector: rentCollector ?? null,
-    configAuthority: configAuthority ?? null,
-    programId: programId ?? PROGRAM_ID,
-  });
-
-  return {
-    instructions: [ix],
-    multisigKey: multisigPda,
-  };
 }
 
 export async function isMultisig(connection: Connection, key: PublicKey) {
@@ -203,17 +167,4 @@ export async function isMultisig(connection: Connection, key: PublicKey) {
   } catch (err) {
     return false;
   }
-}
-
-async function Example() {
-  const connection = new Connection("https://api.mainnet-beta.solana.com");
-  const feePayer = Keypair.generate();
-  const signature = createMultisig({
-    connection,
-    members: createMembers([
-      { key: PublicKey.default, permissions: SquadPermissions.All },
-    ]),
-    creator: PublicKey.default,
-    threshold: 2,
-  });
 }
