+++ sdk/multisig/src/actions/index.ts
@@ -1,9 +1,7 @@
 export * from "./createMultisig";
 export * from "./createVaultTransaction";
 export * from "./createConfigTransaction";
+export * from "./createBatch";
 export * from "./members";
-// WIP
-//export * from "./createTransactionMultiStep";
 
-// WIP
-//export * from "./createBatch";
+export { ConfigActions } from "./common/types";
